+++ app/routes/events/all-events/components/featured-card.component.tsx
@@ -2,7 +2,15 @@ import { Button } from '~/primitives/button/button.primitive';
 import Icon from '~/primitives/icon';
 import HtmlRenderer from '~/primitives/html-renderer';
 import { ContentItemHit } from '~/routes/search/types';
-import { formatEventCardDate } from './all-events-algolia-tree.component';
+
+export function formatEventCardDate(isoDate: string) {
+  return new Date(isoDate).toLocaleDateString('en-US', {
+    timeZone: 'America/New_York',
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+  });
+}
 
 export const FeaturedEventCard = ({ card }: { card: ContentItemHit }) => {
   const {
